Bugfix/pytorch threadlocal issue

* [Bugfix] Reimplementing ThreadLocalPtr with explicit mapping of thread id to object pointers
Previous approach using thread_local specifier was prone to memory leaks: the destructor of ThreadLocalPtr was unable to access all of the created objects

* [Bugfix] More dox

* [Bugfix] Thread-safe storage in ThreadLocalPtr

Approved-by: Philipe Moura

+++ src/backend/tensor.hpp
@@ -136,7 +136,7 @@ class Shape {
     inline bool operator==(const Shape& another) const {
         if (size != another.size)
             return false;
-        for (size_t i = 0; i < size; i++)
+        for (int i = 0; i < size; i++)
             if (shape[i] != another.shape[i])
                 return false;
         return true;
